Added Character object
This makes a distinction between the user's account, which is an
actual, living human, and the character they're playing. A number
of references to the User object have had to be changed to the new
one.

The database has been updated to have the exp and level data. Also,
to store the new character data.