Change a lot of table names
The tables now reflect 'turns', which can have multiple messages in
an order. This gives us the storage capacity to be able to have history
listed about a fight. Before this wasn't possible because we just stored
the messages, impossible to have turn-by-turn playback.

All fight data is now stored, not just the current fight. Stages are
stored on the table, and not worked out by the class.