Made item actions system slightly better. Still working on it though.
Actions can now have parameters, such as "to right hand" for the "Equip"
action.

getActions on an Item now requires the Character who you're interested
in using the item, since a Character can't do certain things (drink
healing potion) in certain situations (when their health is full).

This required rewriting the json that use-item and load-details outputs,
based on a number of changes in the Item class.